+++ src/main/java/com/nextgenpos/nextgenpos/servlets/EditProduct.java
@@ -32,7 +32,7 @@ public class EditProduct extends HttpServlet {
         String productId = request.getParameter("product_id");
         String productName = request.getParameter("product_name");
         Integer quantity = Integer.parseInt(request.getParameter("product_quantity"));
-        Float price = Float.parseFloat(request.getParameter("product_price"));
+        Double price = Double.parseDouble(request.getParameter("product_price"));
         String description = request.getParameter("product_description");
         String provider = request.getParameter("product_provider");
